Code v1.5/requirements.txt
langchain
langsmith
faiss-cpu
python-docx
PyPDF2
mammoth
sentence-transformers
torch
transformers
openai
streamlit
Streamlit-feedback
Langchain-community
python-docx
PyPDF2
Dspy-ai
reportlab